overseer-pixeldiff/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19
Pillow==11.1.0
scikit-image==0.25.0
numpy==2.2.1